Create template.rs
Proposal for template. Includes copyright, contact email, MIT license (to be used when applicable) and authors.
Authors should be updated when people make substantial contributions to the file to be used as a reference for good people to ask questions/code reviews.

B. Marsh is an example name and should be changed as should the year.