Do not double-encode transactions when hashing them
BlakeTwo256::hash_of() already SCALE-encodes the input so calling
hash_of(tx.encode()) doesn't only SCALE-encode the transaction but
also SCALE-encodes the output of tx.encode() which is Vec<u8> meaning
there's two layers of encoding when the hash is calculated.

The call to encode() is unnecessary because hash_of() already
encodes the input to SCALE, it introduces a discrepancy between how
genesis UTXO hash and other UTXO hashes are calculated and complicates
the interaction with the UTXO subsystem for any outside party who wishes
to communicate with Mintlayer core.