UTXO: Adapt input and output formats for more flexibility
This changes the format of transaction inputs and outputs to enable more
future extensibility. For outputs, the `Destination` type is introduced
to differentiate between various spending types such as P2SH, P2PK,
P2PKH and so on. It is an enum so it is easily extended by adding extra
variants. In inputs, there are two new fields: `lock` and `witness`.
These are binary data that are interpreted according to the output
destination.