Utxo: Implement send_to_pubkey()
This function allows the user to provide a public key of destination
and an amount and hides the transaction creation process to make
the UTXO system easier to interact with.

Fix UTXO tests to use H256 as AccountId (like Substrate uses elsewhere)
to allow send_to_pubkey() to derive the caller's public key from it
and use it to sign the UTXO it tries to send.