pools: Implement Contract-to-P2PK transfers
This commit implements transfers from contract to a public key and
thus allows smart contracts to escape the account model to make actual
UTXO transactions.

It implements an initial version of the consensus-critical coin
picking and condensing algorithm and utilizes the OP_SPEND opcode
to spend the OP_CALL vins.

Furthermore, it implements transfer() and balance() methods for
the ChainExtension in pp/lib.rs which allow smart contracts to
transparently interact with the UTXO system.