pools: Implement Contract-to-Contract transfers
Similar to Contract-to-Pubkey, C2C also escapes the account model and
converts the smart contract call into an OP_CALL UTXO and spends all of
the UTXO of the caller smart contract. Unlike with Contract-to-Pubkey,
the vout of this TX is another OP_CALL that calls the other contract
and transfers all of the funds of the caller to that contract.

This commit also implements a new call() ChainExtension method that can
be used to execute generic cross-contract calls over UTXO.