utxo: Improve OP_CALL spending validation
When PP is constructing the final TX that spends the smart contract's
funds, save the outpoint hashes of the vins to a separate storage.

This indicates to the local node that the smart contract has spent
its funds and when it receives a TX trying to spend OP_CALLs of a given
smart contract, the outpoint hashes in the witness field can be queried
from this storage and if they are found, the TX is valid (or at least
the local node agrees with the TX and accepts it).