test: Convert str to bytevector in DestCreatePP::load()
When the UTXO set is queried from the node, it returns the bytecode
of the smart contract as a long string. Previously the code interpreted
strings as filepaths and thus incorrectly tried to open a file using
the bytecode string.

Convert the string representation to a bytevector so that the
DestCreatePP object can be created correctly.