pools: Prevent contracts from calling themselves
Prevent a contract from calling itself through the ChainExtension as
the smart contract already has capabilities of calling its own functions
even without the existence of the ChainExtension, makes it harder to
handle contract balances correctly and makes it very easy for the user
to shoot themselves in the foot by accidentally calling the function
that called into the ChainExtension, effectively resulting in infite
recursion (bounded by the amount of gas, obviously).

If this is found to be needed, the issue can be revisited.