utxo: Implement C2X transfer refunds
Implement Contract-to-Pubkey (C2PK) and Contract-to-Contract (C2C) refunds.
As the contract must be spend all of its funds when it's doing a C2X
transfer, a refunding mechanism where the excess funds of a transfer
are sent back to the contract must be implemented.

This commit uses Destination::FundPP to refund the contract the
difference of total funds versus the amount it is transferring.